updating elogind

+++ PKGBUILD
@@ -11,7 +11,7 @@ arch=('x86_64')
 url="https://github.com/elogind/elogind"
 license=('LGPL-2.1-or-later')
 depends=('dbus' 'pam' 'libcap' 'eudev' 'util-linux')
-makedepends=('meson' 'ninja' 'gperf' 'python-jinja' 'docbook-xsl')
+makedepends=('meson' 'ninja' 'gperf' 'python-jinja')
 provides=('libelogind')
 conflicts=('systemd-libs')
 source=("https://github.com/elogind/elogind/archive/refs/tags/v${pkgver}.tar.gz")
@@ -26,13 +26,17 @@ build() {
         -Dcgroup-controller=elogind \
         -Ddefault-hierarchy=unified \
         -Dpamlibdir=/usr/lib/security \
-        -Dman=true
+        -Dman=false
     meson compile -C build
 }
 
 package() {
     cd "${srcdir}/${pkgname}-${pkgver}"
     DESTDIR="${pkgdir}" meson install -C build
+    if [ ! -f "${pkgdir}/usr/lib/elogind/elogind" ]; then
+        echo "ERROR: /usr/lib/elogind/elogind not found after install! Build/install failed silently." >&2
+        exit 1
+    fi
     ln -s elogind/sd-login.h "${pkgdir}/usr/include/sd-login.h"
     rm -rf "${pkgdir}/usr/lib/systemd"
     rm -rf "${pkgdir}/lib/systemd"
